move back to the org.apache.skywalking prefix, only profile e2e move to test.apache.skywalking
