+++ test/e2e/e2e-service-provider/src/main/java/org/apache/skywalking/e2e/Service0Application.java
@@ -16,9 +16,18 @@
  *
  */
 
-package test.apache.skywalking.e2e;
+package org.apache.skywalking.e2e;
 
-import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.context.annotation.ComponentScan;
+import org.springframework.data.jpa.repository.config.EnableJpaRepositories;
 
-public interface UserRepo extends JpaRepository<User, Long> {
+@EnableJpaRepositories
+@SpringBootApplication
+@ComponentScan({"org.apache.skywalking", "test.apache.skywalking"})
+public class Service0Application {
+    public static void main(String[] args) {
+        SpringApplication.run(Service0Application.class, args);
+    }
 }
