+++ test/e2e/e2e-service-provider/src/main/java/test/apache/skywalking/e2e/profile/CreateUser.java
@@ -19,7 +19,7 @@
 package test.apache.skywalking.e2e.profile;
 
 import lombok.Data;
-import test.apache.skywalking.e2e.User;
+import org.apache.skywalking.e2e.User;
 
 @Data
 public class CreateUser {
