+++ test/e2e/e2e-service-provider/src/main/java/test/apache/skywalking/e2e/profile/ProfileController.java
@@ -20,8 +20,8 @@ package test.apache.skywalking.e2e.profile;
 
 import java.util.concurrent.TimeUnit;
 import lombok.RequiredArgsConstructor;
-import test.apache.skywalking.e2e.User;
-import test.apache.skywalking.e2e.UserRepo;
+import org.apache.skywalking.e2e.User;
+import org.apache.skywalking.e2e.UserRepo;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RestController;
